Allow projects to build a dockerfile on top of a Dock container during extension
Currently during dock extension, if a project requires building a
dockerfile for environment setup outside of what is specified
within its .dock file (e.g. copying files and installing
applications), then setups for previously extended projects in the
same manner will be lost due to each project's Dockerfile being based
on a base image (see FROM <image>) rather than the image representing
the Dock container. This should be the case, since each project can be
brought up in isolation.

To get around this, we assume each project set to be extended into
a dock container has a compatible base with the target dock container
and build the project's dockerfile on top of the image for that dock
container.