Remove building of project development dockerfiles during extension
Project development dockerfiles are mainly used for building the
outer Dock container and installing development tools specific to
that project. This is problematic when Dock containers are being
extended since multiple projects can not only possess different and
perhaps conflicting dev tools but also virtually anything else that
one might see fit to include within the development dockerfile.

With this change, we'll just inspect the dockerfile for its base image
if necessary (omitting the remaining dockerfile instructions) when building
the outer Dock container for extension and leave identifying a solution
which allows for the coexistence of project specific development tools as
future work.