Fix port_taken_on_localhost method for OSX machines
The netstat command that was used previously was not specifically
looking at LISTENING ports and was hitting false positives when there
was an open connection to a remote mysql instance as well as when you
had a port of like 409200 open it would fail the check for publishing
port 9200 (used by elasticsearch).

This lsof check will only show ports that are open and listening and
is explicit to the port being checked.